Fix errors during the documentation build in Read the docs (#61)
Solves the following issues mentioned in #57.

The documentation should be accessible at https://pysteps.readthedocs.io/en/latest and https://pysteps.readthedocs.io

Solve issue with the missing vet module. Now, the pysteps package is installed in a virtual environment during the build in Read the docs and that installed package is used to build the documentation.