Automate pytests with tox (#106)
Automate tests, and other tasks using tox.

Requirements: conda, tox, and tox-conda installed.

Tox is used to automate tasks by just executing the tox command, without the need to set up the right development environment or downloading the pysteps-data beforehand.

From any environment with tox installed, the tests can be run just executing the tox command from the project root (where the new tox.ini file is located).
This will create a virtual environment with all the necessary dependencies to test pysteps and also download the pysteps data automatically.

The tox-conda plugin was added to create the environments using conda. This was necessary to install basemaps and cartopy without major complications.

tox environments:
- default (testenv): Run the pysteps's test suite
- install: Test the installation of the package in an environment with the minimal dependencies (run  
  with -r option)
- pypi: Test the installation of the package from the PyPI in a clean environment (run with -r option)
- pypi_test: Test the installation of the package from the test-PyPI in a clean environment (run with    
  -r option)
- pypi_test_full: Test the installation of the package from the test-PyPI in an environment with all the  
  dependencies (run with -r option)
- black: run black tests
- docs: Build the html documentation using sphinx
- install: Test the installation of the package in an environment with the minimal dependencies
- pypi: Test the installation of the package from the PyPI in a clean environment
- pypi_test: Test the installation of the package from the test-PyPI in a clean environment

The environments can be run with tox -e <environment_name>. By default, the environment are created only when the specifications in the tox.ini file changes. 
To force the recreation of the environment, add -r (recommended for some testing environments).

The documentation was updated to include the tox usage.