Refactor download_mrms_data
The MRMS example data now is part of the pysteps-data repository.
Hence, the download_mrms_data function is not used anymore to install
the pysteps data.
The function was refactored for the more general purpose of downloading
data from the mtarchive.geol.iastate.edu over a requested range of dates.