Revert addition of MRMS example :(
The new MRMS example added in the previous commits needs pygrib
and, optionally, cartopy.
When Pygrib is installed using pip in RTD, the grib libraries are not
installed, and the documentation build fails.

Using conda to manage the packages in RTD exceeded the memory limit of
1GB, and the build also fails.

Since the solution of this issue is not trivial. Better revert the
changes and include the MRMS example in the future.