Update thunderstorm_detection_and_tracking.py
I added some extra comment lines and I made some minor changes to the final two plots. Another idea, but this may just be looking for more work: would it be possible to link the track colours in the first figure to the colour of the tracks in the two evaluation figures? Other than that, nice example! :)